Rename to ShareMenuContainer
